Fix: Infinite page reload in Matches component
The Matches component was stuck in an infinite reload loop if there were
no results for the match. This checks to see if there are zero results
and if so prevents the reload.

+++ src/components/Matches.jsx
@@ -67,7 +67,7 @@ export default function Matches({ input, searchAlgorithm, matches, precomputed }
 
   const requiredResultCount = (page * matchesPerPage) - (matchesPerPage - 1);
 
-  if (matches?.results?.length < requiredResultCount) {
+  if (matches?.results?.length > 0 && matches?.results?.length < requiredResultCount) {
     query.delete('page');
     location.search = query.toString();
     history.replace(location);
